+++ LifeSimulator/app/src/main/java/com/example/lifesimulator/Model/Item.java
@@ -4,7 +4,7 @@ public class Item extends NetWorth{
     public Item() {
     }
 
-    public Item(String code, String name, float price, String description, boolean isOwner) {
+    public Item(String code, String name, int price, String description, boolean isOwner) {
         super(code, name, price, description, isOwner);
     }
 }
